A custom model from a custom module
