+++ app/pages.py
@@ -3,11 +3,15 @@
 
 """ Main routes """
 
+import importlib
 from flask import Blueprint, \
     render_template, request, flash, redirect, url_for
 from app import forms
 from config import user_config
+from .basemodel import db, model2table
+from .forms import module
 
+MyModel = module.MyModel
 blueprint = Blueprint('pages', __name__)
 
 # ######################################################
@@ -29,26 +33,25 @@ blueprint = Blueprint('pages', __name__)
 #     return decorator
 
 ######################################################
-## PAOLO
-######################################################
+MyTable = model2table(MyModel)
 
-from .models import db, MyModel, MyTable
 
 def insertdb(iform, obj):
     iform.populate_obj(obj)
-    #flash("Populated user %s" % dir(user), 'success')
+    # flash("Populated user %s" % dir(user), 'success')
     db.session.add(obj)
-    # Save into db
+    # Save into db
     db.session.commit()
+
 
 def row2dict(r):
     """ Convert a single sqlalchemy row into a dictionary """
-    #http://stackoverflow.com/a/1960546/2114395
+    # http://stackoverflow.com/a/1960546/2114395
     return {c.name: str(getattr(r, c.name)) for c in r.__table__.columns}
 
 template = 'forms/insert_search.html'
 
-#@cached
+
 @blueprint.route('/view', methods=["GET", "POST"])
 def view():
     status = "View"
@@ -61,7 +64,7 @@ def view():
         from sqlalchemy import desc
         field = desc(field)
 
-    # SQLalchemy query (sorted)
+    # SQLalchemy query (sorted)
     data = MyModel.query.order_by(field)
 
     items = []
@@ -72,6 +75,7 @@ def view():
         table=MyTable(items, sort_by=sort_field, sort_reverse=reverse),
         status=status, formname='view',
         **user_config['content'])
+
 
 @blueprint.route('/insert', methods=["GET", "POST"])
 def insert():
@@ -87,24 +91,22 @@ def insert():
         status=status, form=iform, formname='insert',
         **user_config['content'])
 
+
 @blueprint.route('/search', methods=["GET", "POST"])
 def search():
     status = "Waiting data to search"
     iform = forms.UserForm()
     if iform.validate_on_submit():
         status = "Work in progress"
-        #flash("User saved", 'success')
+        # flash("User saved", 'success')
 
     return render_template(template,
         status=status, form=iform, formname='search',
         **user_config['content'])
 
-######################################################
-######################################################
-
 
 ################
-#### basic interface routes ####
+# Basic interface routes ####
 ################
 
 @blueprint.route('/')
