+++ config/base/models.json
@@ -1,3 +1,3 @@
 {
-	name: 'models'
+	"model": "example"
 }
