+++ config/base/options.json
@@ -1,3 +1,3 @@
 {
-	name: 'opt'
+	"options": "some"
 }
